+++ java-vtl-script/src/main/java/no/ssb/vtl/script/visitors/ParamVisitor.java
@@ -24,7 +24,7 @@ public class ParamVisitor extends VTLBaseVisitor<Object> {
         } else if (ctx.DOUBLE_CONSTANT() != null) {
             return Double.valueOf(constant);
         } else if (ctx.INTEGER_CONSTANT() != null) {
-            return Integer.valueOf(constant);
+            return Long.valueOf(constant);
         } else if (ctx.NULL_CONSTANT() != null) {
             return null;
         } else { //String
