+++ java-vtl-script/src/test/java/no/ssb/vtl/script/operations/hierarchy/HierarchyOperationTest.java
@@ -56,34 +56,34 @@ public class HierarchyOperationTest extends RandomizedTest {
 
     private static List<Object> createAggregatedPopulation() {
         return Lists.newArrayList(
-                Year.of(2000), "Austria", 2000, -2000L,
-                Year.of(2000), "Belgium", 2000, -2000L,
-                Year.of(2000), "European Union", 10000, -10000L,
-                Year.of(2000), "Luxembourg", 2000, -2000L,
-                Year.of(2000), "Benelux", 6000, -6000L,
-                Year.of(2000), "Italy", 2000, -2000L,
-                Year.of(2000), "Holland", 2000, -2000L,
-                Year.of(2001), "Austria", 2001, -2001L,
-                Year.of(2001), "Belgium", 2001, -2001L,
-                Year.of(2001), "European Union", 10005, -10005L,
-                Year.of(2001), "Luxembourg", 2001, -2001L,
-                Year.of(2001), "Benelux", 6003, -6003L,
-                Year.of(2001), "Italy", 2001, -2001L,
-                Year.of(2001), "Holland", 2001, -2001L,
-                Year.of(2002), "Austria", 2002, -2002L,
-                Year.of(2002), "Belgium", 2002, -2002L,
-                Year.of(2002), "European Union", 10010, -10010L,
-                Year.of(2002), "Luxembourg", 2002, -2002L,
-                Year.of(2002), "Benelux", 6006, -6006L,
-                Year.of(2002), "Italy", 2002, -2002L,
-                Year.of(2002), "Holland", 2002, -2002L,
-                Year.of(2003), "Austria", 2003, -2003L,
-                Year.of(2003), "Belgium", 2003, -2003L,
-                Year.of(2003), "European Union", 10015, -10015L,
-                Year.of(2003), "Luxembourg", 2003, -2003L,
-                Year.of(2003), "Benelux", 6009, -6009L,
-                Year.of(2003), "Italy", 2003, -2003L,
-                Year.of(2003), "Holland", 2003, -2003L
+                Year.of(2000), "Austria", 2000L, -2000L,
+                Year.of(2000), "Belgium", 2000L, -2000L,
+                Year.of(2000), "European Union", 10000L, -10000L,
+                Year.of(2000), "Luxembourg", 2000L, -2000L,
+                Year.of(2000), "Benelux", 6000L, -6000L,
+                Year.of(2000), "Italy", 2000L, -2000L,
+                Year.of(2000), "Holland", 2000L, -2000L,
+                Year.of(2001), "Austria", 2001L, -2001L,
+                Year.of(2001), "Belgium", 2001L, -2001L,
+                Year.of(2001), "European Union", 10005L, -10005L,
+                Year.of(2001), "Luxembourg", 2001L, -2001L,
+                Year.of(2001), "Benelux", 6003L, -6003L,
+                Year.of(2001), "Italy", 2001L, -2001L,
+                Year.of(2001), "Holland", 2001L, -2001L,
+                Year.of(2002), "Austria", 2002L, -2002L,
+                Year.of(2002), "Belgium", 2002L, -2002L,
+                Year.of(2002), "European Union", 10010L, -10010L,
+                Year.of(2002), "Luxembourg", 2002L, -2002L,
+                Year.of(2002), "Benelux", 6006L, -6006L,
+                Year.of(2002), "Italy", 2002L, -2002L,
+                Year.of(2002), "Holland", 2002L, -2002L,
+                Year.of(2003), "Austria", 2003L, -2003L,
+                Year.of(2003), "Belgium", 2003L, -2003L,
+                Year.of(2003), "European Union", 10015L, -10015L,
+                Year.of(2003), "Luxembourg", 2003L, -2003L,
+                Year.of(2003), "Benelux", 6009L, -6009L,
+                Year.of(2003), "Italy", 2003L, -2003L,
+                Year.of(2003), "Holland", 2003L, -2003L
         );
     }
 
@@ -125,7 +125,7 @@ public class HierarchyOperationTest extends RandomizedTest {
                 .put("FUNKSJON_KAPITTEL", IDENTIFIER, String.class)
                 .put("KONTOKLASSE", IDENTIFIER, String.class)
                 .put("REGION", IDENTIFIER, String.class)
-                .put("BELOP", MEASURE, Integer.class)
+                .put("BELOP", MEASURE, Long.class)
                 .build();
     }
 
@@ -186,7 +186,7 @@ public class HierarchyOperationTest extends RandomizedTest {
         DataStructure structure = DataStructure.builder()
                 .put("Year", IDENTIFIER, Year.class)
                 .put("Country", IDENTIFIER, String.class)
-                .put("Population", MEASURE, Integer.class)
+                .put("Population", MEASURE, Long.class)
                 .put("OtherPopulation", MEASURE, Long.class)
                 .build();
 
@@ -211,7 +211,7 @@ public class HierarchyOperationTest extends RandomizedTest {
                 DataPoint point = structure.wrap(ImmutableMap.of(
                         "Year", year,
                         "Country", country,
-                        "Population", year.getValue(), //randomIntBetween(0, 20)
+                        "Population", (long) year.getValue(), //randomIntBetween(0, 20)
                         "OtherPopulation", (long) year.getValue() * -1
                 ));
                 data.add(point);
@@ -328,13 +328,13 @@ public class HierarchyOperationTest extends RandomizedTest {
         DataStructure structure = DataStructure.builder()
                 .put("id1", IDENTIFIER, String.class)
                 .put("id2", IDENTIFIER, String.class)
-                .put("m1", MEASURE, Integer.class)
+                .put("m1", MEASURE, Long.class)
                 .build();
 
         DataStructure otherStructure = DataStructure.builder()
                 .put("id1", IDENTIFIER, String.class)
                 .put("id2", IDENTIFIER, String.class)
-                .put("m1", MEASURE, Integer.class)
+                .put("m1", MEASURE, Long.class)
                 .build();
 
         Dataset dataset = createEmptyDataset(structure);
@@ -351,7 +351,7 @@ public class HierarchyOperationTest extends RandomizedTest {
         DataStructure structure = DataStructure.builder()
                 .put("id1", IDENTIFIER, String.class)
                 .put("id2", IDENTIFIER, String.class)
-                .put("m1", MEASURE, Integer.class)
+                .put("m1", MEASURE, Long.class)
                 .build();
 
         Dataset dataset = createEmptyDataset(structure);
@@ -368,7 +368,7 @@ public class HierarchyOperationTest extends RandomizedTest {
         DataStructure structure = DataStructure.builder()
                 .put("id1", IDENTIFIER, String.class)
                 .put("id2", IDENTIFIER, String.class)
-                .put("m1", MEASURE, Integer.class)
+                .put("m1", MEASURE, Long.class)
                 .put("m2", MEASURE, Double.class)
                 .put("m3", MEASURE, Float.class)
                 .put("m4", MEASURE, String.class)
